net volume error

+++ 02.Firmware/src/inc/hardware.h
@@ -395,8 +395,8 @@ typedef struct
     float fw_flow_max;
     float reserve_flow_min;
     float reserve_flow_max;
-    float fw_flow_first_counter;
-    float reverse_flow_first_counter;
+    uint32_t fw_flow_first_counter;
+    uint32_t reverse_flow_first_counter;
     float fw_flow_sum;
     float reverse_flow_sum;
     uint8_t valid;
